requirements.txt
# Minimal requirements for Vercel deployment
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
SQLAlchemy==1.4.53
psycopg2-binary==2.9.9

# Authentication
passlib==1.7.4
bcrypt==4.3.0
python-jose==3.3.0
cryptography==45.0.7
authlib==1.3.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0

# HTTP
httpx==0.25.2
requests==2.32.3
aiohttp==3.9.1
aiofiles==23.2.1

# File handling
python-multipart==0.0.6
python-dotenv==1.0.0
PyYAML==6.0.2

# Basic processing
numpy==1.26.4
pillow==11.3.0
pydub==0.25.1

# Utilities
click==8.2.1
packaging==25.0
typing_extensions==4.12.2

# Payment
stripe==7.8.0
